Рефакторинг управления ролями и разрешениями пользователей: добавление модели ролей, событий и сопутствующих функций
