+++ alembic/versions/6b6908f826ac_initial.py
@@ -1,0 +1,226 @@
+"""initial
+
+Revision ID: 6b6908f826ac
+Revises:
+Create Date: 2026-01-27 19:41:17.201396
+
+"""
+
+from collections.abc import Sequence
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "6b6908f826ac"
+down_revision: str | Sequence[str] | None = None
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "achievements",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("icon_url", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "competencies",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "levels",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("level_type", sa.String(length=50), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("required_points", sa.Integer(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "roles",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("name"),
+    )
+    op.create_table(
+        "task_solution_statuses",
+        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "user_activity_statuses",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "users",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("first_name", sa.Text(), nullable=False),
+        sa.Column("last_name", sa.Text(), nullable=True),
+        sa.Column("patronymic", sa.Text(), nullable=True),
+        sa.Column("phone_number", sa.Text(), nullable=True),
+        sa.Column("telegram_id", sa.BigInteger(), nullable=False),
+        sa.Column("github_url", sa.Text(), nullable=True),
+        sa.Column("join_date", sa.Date(), server_default=sa.text("(CURRENT_DATE)"), nullable=False),
+        sa.Column("activity_status_id", sa.BigInteger(), nullable=True),
+        sa.Column("academic_points", sa.Integer(), nullable=False),
+        sa.Column("reputation_points", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["activity_status_id"],
+            ["user_activity_statuses.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("telegram_id"),
+    )
+    op.create_table(
+        "role_events",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("to_user_id", sa.Integer(), nullable=False),
+        sa.Column("from_user_id", sa.Integer(), nullable=False),
+        sa.Column(
+            "operand",
+            sa.Enum(
+                "ADD", "DELETE", "REPLACE", name="role_event_operand_enum", native_enum=False, create_constraint=True
+            ),
+            nullable=False,
+        ),
+        sa.Column("role_id", sa.Integer(), nullable=False),
+        sa.Column("reason", sa.Text(), nullable=True),
+        sa.Column("created_at", sa.DateTime(), server_default=sa.text("(CURRENT_TIMESTAMP)"), nullable=False),
+        sa.ForeignKeyConstraint(["from_user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["role_id"], ["roles.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["to_user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "tasks",
+        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=False),
+        sa.Column("created_date", sa.Date(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.Column("due_date", sa.Date(), nullable=False),
+        sa.Column("user_id", sa.BigInteger(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "user_achievements",
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("achievements_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["achievements_id"], ["achievements.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("user_id", "achievements_id"),
+    )
+    op.create_table(
+        "user_competencies",
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("competence_id", sa.Integer(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(["competence_id"], ["competencies.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("user_id", "competence_id"),
+    )
+    op.create_table(
+        "user_levels",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("level_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["level_id"],
+            ["levels.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "level_id", name="uq_user_level"),
+    )
+    op.create_table(
+        "user_roles",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("role_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["role_id"], ["roles.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "role_id", name="uq_user_role"),
+    )
+    op.create_table(
+        "valuations",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("recipient_id", sa.BigInteger(), nullable=False),
+        sa.Column("giver_id", sa.BigInteger(), nullable=False),
+        sa.Column("reason", sa.Text(), nullable=True),
+        sa.Column("points", sa.Integer(), nullable=False),
+        sa.Column("points_type", sa.String(length=50), nullable=False),
+        sa.Column("created_at", sa.DateTime(), server_default=sa.text("(CURRENT_TIMESTAMP)"), nullable=False),
+        sa.ForeignKeyConstraint(["giver_id"], ["users.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["recipient_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "task_solutions",
+        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
+        sa.Column("task_id", sa.BigInteger(), nullable=False),
+        sa.Column("user_id", sa.BigInteger(), nullable=False),
+        sa.Column("status_id", sa.BigInteger(), nullable=False),
+        sa.Column("content", sa.Text(), nullable=False),
+        sa.Column("solution_url", sa.Text(), nullable=False),
+        sa.Column("created_date", sa.Date(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["status_id"],
+            ["task_solution_statuses.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["task_id"],
+            ["tasks.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("task_solutions")
+    op.drop_table("valuations")
+    op.drop_table("user_roles")
+    op.drop_table("user_levels")
+    op.drop_table("user_competencies")
+    op.drop_table("user_achievements")
+    op.drop_table("tasks")
+    op.drop_table("role_events")
+    op.drop_table("users")
+    op.drop_table("user_activity_statuses")
+    op.drop_table("task_solution_statuses")
+    op.drop_table("roles")
+    op.drop_table("levels")
+    op.drop_table("competencies")
+    op.drop_table("achievements")
+    # ### end Alembic commands ###
