--- alembic/versions/72cc8f0a976f_initialize_database_with_sqlite_support.py
@@ -1,421 +1,0 @@
-"""Initialize database with SQLite support
-
-Revision ID: 72cc8f0a976f
-Revises:
-Create Date: 2026-01-18 21:20:18.497847
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "72cc8f0a976f"
-down_revision: str | Sequence[str] | None = None
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "academic_roles",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "achievements",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("icon_url", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "admin_roles",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "attachments_types",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "competencies",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "levels",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("level_type", sa.String(length=50), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("required_points", sa.Integer(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "project_member_roles",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "projects_statuses",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "task_solution_statuses",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "user_activity_statuses",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "attachments",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("content", sa.Text(), nullable=False),
-        sa.Column("type_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["type_id"],
-            ["attachments_types.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "users",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("first_name", sa.Text(), nullable=False),
-        sa.Column("last_name", sa.Text(), nullable=True),
-        sa.Column("patronymic", sa.Text(), nullable=True),
-        sa.Column("phone_number", sa.Text(), nullable=True),
-        sa.Column("telegram_id", sa.BigInteger(), nullable=False),
-        sa.Column("github_url", sa.Text(), nullable=True),
-        sa.Column("join_date", sa.Date(), server_default=sa.text("(CURRENT_DATE)"), nullable=False),
-        sa.Column("activity_status_id", sa.BigInteger(), nullable=True),
-        sa.Column("academic_points", sa.Integer(), nullable=False),
-        sa.Column("reputation_points", sa.Integer(), nullable=False),
-        sa.Column("academic_role_id", sa.BigInteger(), nullable=True),
-        sa.Column("admin_role_id", sa.BigInteger(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["academic_role_id"],
-            ["academic_roles.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["activity_status_id"],
-            ["user_activity_statuses.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["admin_role_id"],
-            ["admin_roles.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("telegram_id"),
-    )
-    op.create_table(
-        "comments",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("user_id", sa.Integer(), nullable=False),
-        sa.Column("content", sa.Text(), nullable=False),
-        sa.Column("datetime", sa.TIMESTAMP(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "projects",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.Column("created_date", sa.Date(), nullable=False),
-        sa.Column("status_id", sa.BigInteger(), nullable=False),
-        sa.Column("user_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["status_id"],
-            ["projects_statuses.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "tasks",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.Column("created_date", sa.Date(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.Column("due_date", sa.Date(), nullable=False),
-        sa.Column("user_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "user_achievements",
-        sa.Column("user_id", sa.Integer(), nullable=False),
-        sa.Column("achievements_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(["achievements_id"], ["achievements.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("user_id", "achievements_id"),
-    )
-    op.create_table(
-        "user_competencies",
-        sa.Column("user_id", sa.Integer(), nullable=False),
-        sa.Column("competence_id", sa.Integer(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.ForeignKeyConstraint(["competence_id"], ["competencies.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("user_id", "competence_id"),
-    )
-    op.create_table(
-        "user_levels",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("user_id", sa.Integer(), nullable=False),
-        sa.Column("level_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["level_id"],
-            ["levels.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("user_id", "level_id", name="uq_user_level"),
-    )
-    op.create_table(
-        "valuations",
-        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
-        sa.Column("recipient_id", sa.BigInteger(), nullable=False),
-        sa.Column("giver_id", sa.BigInteger(), nullable=False),
-        sa.Column("reason", sa.Text(), nullable=True),
-        sa.Column("points", sa.Integer(), nullable=False),
-        sa.Column("points_type", sa.String(length=50), nullable=False),
-        sa.Column("created_at", sa.DateTime(), server_default=sa.text("(CURRENT_TIMESTAMP)"), nullable=False),
-        sa.ForeignKeyConstraint(["giver_id"], ["users.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["recipient_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "comment_answers",
-        sa.Column("comment_id", sa.Integer(), nullable=False),
-        sa.Column("answer_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["answer_id"],
-            ["comments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["comment_id"],
-            ["comments.id"],
-        ),
-        sa.PrimaryKeyConstraint("comment_id", "answer_id"),
-    )
-    op.create_table(
-        "comment_attachments",
-        sa.Column("comment_id", sa.Integer(), nullable=False),
-        sa.Column("attachments_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["attachments_id"],
-            ["attachments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["comment_id"],
-            ["comments.id"],
-        ),
-        sa.PrimaryKeyConstraint("comment_id", "attachments_id"),
-    )
-    op.create_table(
-        "project_achievements",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("project_id", sa.BigInteger(), nullable=False),
-        sa.Column("achievements_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["achievements_id"],
-            ["achievements.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["project_id"],
-            ["projects.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "project_attachments",
-        sa.Column("project_id", sa.BigInteger(), nullable=False),
-        sa.Column("attachments_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["attachments_id"],
-            ["attachments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["project_id"],
-            ["projects.id"],
-        ),
-        sa.PrimaryKeyConstraint("project_id", "attachments_id"),
-    )
-    op.create_table(
-        "project_comments",
-        sa.Column("project_id", sa.BigInteger(), nullable=False),
-        sa.Column("comment_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["comment_id"],
-            ["comments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["project_id"],
-            ["projects.id"],
-        ),
-        sa.PrimaryKeyConstraint("project_id", "comment_id"),
-    )
-    op.create_table(
-        "project_members",
-        sa.Column("project_id", sa.BigInteger(), nullable=False),
-        sa.Column("user_id", sa.BigInteger(), nullable=False),
-        sa.Column("role_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["project_id"],
-            ["projects.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["role_id"],
-            ["project_member_roles.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("project_id", "user_id"),
-    )
-    op.create_table(
-        "task_attachments",
-        sa.Column("task_id", sa.BigInteger(), nullable=False),
-        sa.Column("attachments_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["attachments_id"],
-            ["attachments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["task_id"],
-            ["tasks.id"],
-        ),
-        sa.PrimaryKeyConstraint("task_id", "attachments_id"),
-    )
-    op.create_table(
-        "task_comments",
-        sa.Column("task_id", sa.BigInteger(), nullable=False),
-        sa.Column("comment_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["comment_id"],
-            ["comments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["task_id"],
-            ["tasks.id"],
-        ),
-        sa.PrimaryKeyConstraint("task_id", "comment_id"),
-    )
-    op.create_table(
-        "task_solutions",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("task_id", sa.BigInteger(), nullable=False),
-        sa.Column("user_id", sa.BigInteger(), nullable=False),
-        sa.Column("status_id", sa.BigInteger(), nullable=False),
-        sa.Column("content", sa.Text(), nullable=False),
-        sa.Column("solution_url", sa.Text(), nullable=False),
-        sa.Column("created_date", sa.Date(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["status_id"],
-            ["task_solution_statuses.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["task_id"],
-            ["tasks.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "task_solution_comments",
-        sa.Column("task_solutions_id", sa.BigInteger(), nullable=False),
-        sa.Column("comment_id", sa.BigInteger(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["comment_id"],
-            ["comments.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["task_solutions_id"],
-            ["task_solutions.id"],
-        ),
-        sa.PrimaryKeyConstraint("task_solutions_id", "comment_id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("task_solution_comments")
-    op.drop_table("task_solutions")
-    op.drop_table("task_comments")
-    op.drop_table("task_attachments")
-    op.drop_table("project_members")
-    op.drop_table("project_comments")
-    op.drop_table("project_attachments")
-    op.drop_table("project_achievements")
-    op.drop_table("comment_attachments")
-    op.drop_table("comment_answers")
-    op.drop_table("valuations")
-    op.drop_table("user_levels")
-    op.drop_table("user_competencies")
-    op.drop_table("user_achievements")
-    op.drop_table("tasks")
-    op.drop_table("projects")
-    op.drop_table("comments")
-    op.drop_table("users")
-    op.drop_table("attachments")
-    op.drop_table("user_activity_statuses")
-    op.drop_table("task_solution_statuses")
-    op.drop_table("projects_statuses")
-    op.drop_table("project_member_roles")
-    op.drop_table("levels")
-    op.drop_table("competencies")
-    op.drop_table("attachments_types")
-    op.drop_table("admin_roles")
-    op.drop_table("achievements")
-    op.drop_table("academic_roles")
-    # ### end Alembic commands ###
