+++ src/pybot/core/constants.py
@@ -4,3 +4,15 @@ from enum import StrEnum
 class PointsTypeEnum(StrEnum):
     ACADEMIC = "academic"
     REPUTATION = "reputation"
+
+
+class RoleEventOperandEnum(StrEnum):
+    ADD = "add"
+    DELETE = "delete"
+    REPLACE = "replace"
+
+
+class RoleEnum(StrEnum):
+    STUDENT = "Student"
+    MENTOR = "Mentor"
+    ADMIN = "Admin"
