+++ src/pybot/db/models/role_module/__init__.py
@@ -1,0 +1,12 @@
+from ...base_class import Base
+
+from .roles import Role
+from .user_roles import UserRole
+from .role_events import RoleEvent
+
+__all__ = [
+    "Base",
+    "Role",
+    "UserRole",
+    "RoleEvent",
+]
