--- src/pybot/db/models/user_module/academic_role.py
@@ -1,20 +1,0 @@
-from __future__ import annotations
-
-from sqlalchemy import Integer, Text
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-from ...base_class import Base
-from .user import User
-
-
-class AcademicRole(Base):
-    __tablename__ = "academic_roles"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    name: Mapped[str] = mapped_column(Text, nullable=False)
-    description: Mapped[str | None] = mapped_column(Text)
-
-    users: Mapped[list[User]] = relationship(
-        "User",
-        back_populates="academic_role",
-    )
