+++ src/pybot/db/models/user_module/user_level.py
@@ -7,12 +7,11 @@ from . import Level, User
 
 class UserLevel(Base):
     __tablename__ = "user_levels"
+    __table_args__ = (UniqueConstraint("user_id", "level_id", name="uq_user_level"),)
 
     id: Mapped[int] = mapped_column(Integer, autoincrement=True, primary_key=True)
     user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
     level_id: Mapped[int] = mapped_column(Integer, ForeignKey("levels.id"), nullable=False)
-
-    __table_args__ = (UniqueConstraint("user_id", "level_id", name="uq_user_level"),)
 
     user: Mapped[User] = relationship("User", back_populates="user_levels")
     level: Mapped[Level] = relationship("Level", back_populates="user_levels")
